Fix x5EC type instead of casting (#3447)
(cherry picked from commit 1626556005e04a9fa11c309bf5b5ce62e6c5599d)

+++ src/melee/ft/chara/ftCommon/ftCo_Damage.c
@@ -505,7 +505,7 @@ block_67:
     if (!(scaled_kb >= p_ftCommonData->x5E8)) {
         goto block_70;
     }
-    ftCommon_8007EFC0(fp, (u32) p_ftCommonData->x5EC);
+    ftCommon_8007EFC0(fp, p_ftCommonData->x5EC);
 block_70:
     if (var_r27 == 0) {
         goto block_75;
